+++ src/translations/en/neo.php
@@ -23,6 +23,8 @@ return [
     'Handle' => 'Handle',
     'How you’ll refer to this block type in the templates.' => 'How you’ll refer to this block type in the templates.',
     'Description' => 'Description',
+    'Icon' => 'Icon',
+    'The SVG icon to use for this block type, if Neo is set to use visual buttons to add new blocks.' => 'The SVG icon to use for this block type, if Neo is set to use visual buttons to add new blocks.',
     'Enabled' => 'Enabled',
     'Whether this block type is allowed to be used.' => 'Whether this block type is allowed to be used.',
     'Max Blocks' => 'Max Blocks',